Simplify testing code
Simplify code for validating and comparing by splitting it up into two
modules, making less things happen automagically, only support
comparisons of flat objects.
